chore: move RoleField component to registry molecules
